SQL-Python-Ecommerce_Project-main/requirement.txt
Flask
pandas
sqlalchemy
mysql-connector-python
